Introduce the CoreRecord defining Object
This class will be used to construct framework defined classes that will
be used to set up a prototype chain between the user's model and the
core framework model. This will utilize prototypical inheritance
strengths of the language rather than the copy pasting of methods that
Rails does to generate attribute/relation based methods.